Feat i16

See merge request swm-13/13_swm07/Oki!21

+++ back-end/src/main/java/com/example/oki/Board/controller/BoardController.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
+import java.util.Optional;
 
 @RequestMapping("/board")
 public class BoardController {
@@ -27,9 +29,14 @@ public class BoardController {
     // 게시판 목록
     @GetMapping
     @ResponseBody
-    public ResponseEntity<Message> list(@RequestParam(value = "keyword") String keyword) throws JsonProcessingException {
+    public ResponseEntity<Message> list(@RequestParam(value = "keyword") String keyword) {
 
         List<Board> boards = boardService.getBoards(keyword);
+
+        if (boards == null) {
+            Message message = new Message(HttpStatus.NOT_FOUND, "Incorrect keyword",  null);
+            return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
+        }
 
         Message message = new Message(HttpStatus.OK, "Success", boards);
 
@@ -40,8 +47,12 @@ public class BoardController {
     @PostMapping
     @ResponseBody
     public ResponseEntity<Message> create(@RequestBody BoardDto boardDto) {
-        boardService.createBoard(boardDto);
+        Long boardId = boardService.createBoard(boardDto);
 
+        if (boardId == null) {
+            Message message = new Message(HttpStatus.NOT_FOUND, "Incorrect user or keyword", null);
+            return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
+        }
         Message message = new Message(HttpStatus.OK, "Success", null);
 
         return new ResponseEntity<>(message, HttpStatus.OK);
@@ -50,9 +61,14 @@ public class BoardController {
     // 게시판 삭제
     @DeleteMapping
     public ResponseEntity<Message> delete(@RequestParam(value = "id") Long id) {
-        Board board = boardService.findBoard(id).get();
+        Optional<Board> board = boardService.findBoard(id);
 
-        String subject = board.getSubject().getKeyword();
+        if (board.isEmpty()) {
+            Message message = new Message(HttpStatus.NOT_FOUND, "Incorrect board", null);
+            return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
+        }
+
+        String subject = board.get().getSubject().getKeyword();
 
         boardService.deleteBoard(id);
